Extract normalizeOptions from index.js
